+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.3.0 / 2014-06-29
 ------------------
 - bugfix: if private key was less than 32 bytes, pad out to 32 bytes with leading zeros (this happens in derive)
-
+- changed behavior of `privateExtendedKey()` and `publicExtendedKey()` to return base 58 encoded `string` instead of `Buffer`
 
 0.2.0 / 2014-06-25
 ------------------